// apps/api/src/modules/client/dto/create-client.dto.d.ts
export declare class CreateClientDto {
    firstName: string;
    lastName: string;
    phone: string;
    alternatePhone?: string;
    email?: string;
    gender?: string;
    dateOfBirth?: string;
    address?: string;
    city?: string;
    district?: string;
    state?: string;
    pincode?: string;
    idProofType?: string;
    idProofNumber?: string;
    customerType?: string;
    birthdayOffer?: boolean;
    notes?: string;
    branchId: string;
}
//# sourceMappingURL=create-client.dto.d.ts.map