// apps/api/src/modules/client/dto/update-client.dto.d.ts
import { CreateClientDto } from './create-client.dto';
declare const UpdateClientDto_base: import("@nestjs/common").Type<Partial<CreateClientDto>>;
export declare class UpdateClientDto extends UpdateClientDto_base {
    isActive?: boolean;
}
export {};
//# sourceMappingURL=update-client.dto.d.ts.map